pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "entrance-monitor"
version = "0.1.0"
description = "Edge-first entrance flow and busyness monitoring prototype for Raspberry Pi 5"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Project Team" }
]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "numpy>=1.26.0,<2.0.0",
  "opencv-python>=4.10.0,<5.0.0",
  "psutil>=6.1.0,<7.0.0",
]

[project.optional-dependencies]
yolo = [
  "ultralytics>=8.3.0,<9.0.0",
  "onnxruntime>=1.20.0,<2.0.0",
]
serial = [
  "pyserial>=3.5,<4.0.0",
]
dev = [
  "pytest>=8.3.0,<9.0.0",
]

[project.scripts]
entrance-monitor = "entrance_monitor.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/entrance_monitor"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
